Separated GoogleMaps management to GoogleMapsManager
